Split test/script/test_script.py by module and fold in test_script_is_new
test_script.py bundled the parser/expression engine tests with a handful
of picard.script.__init__ documentation-helper tests
(script_function_documentation / _all). Split it so each file maps to one
module:

- test/script/test_script.py -> test_parser.py (ScriptParserTest, the
  parser and script-function engine)
- new test/script/test_init.py (ScriptFunctionDocumentationTest, covering
  the script_function_documentation helpers in picard.script.__init__)

Fold test_script_is_new.py (the $get_new / $get_original ScriptParser
tests) into test_parser.py, since it also targets picard.script.parser,
and drop the redundant script_ prefix.

No test logic changed; per-file test method counts reconcile with the
originals (150 + 11 = 161 = 154 in test_parser + 7 in test_init).
