+++ test/script/test_init.py
@@ -1,0 +1,162 @@
+# Picard, the next-generation MusicBrainz tagger
+#
+# Copyright (C) 2007 Lukáš Lalinský
+# Copyright (C) 2010, 2014, 2018-2022, 2024-2026 Philipp Wolfer
+# Copyright (C) 2012 Chad Wilson
+# Copyright (C) 2013 Michael Wiencek
+# Copyright (C) 2013, 2017-2022, 2024-2026 Laurent Monin
+# Copyright (C) 2014, 2017 Sophist-UK
+# Copyright (C) 2016-2017 Sambhav Kothari
+# Copyright (C) 2017 Antonio Larrosa
+# Copyright (C) 2017-2018, 2021 Wieland Hoffmann
+# Copyright (C) 2018 virusMac
+# Copyright (C) 2020-2022, 2025 Bob Swift
+# Copyright (C) 2021 Adam James
+# Copyright (C) 2025 Thuna-Cing
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, see <https://www.gnu.org/licenses/>.
+
+
+import unittest
+from unittest.mock import patch
+
+from test.picardtestcase import PicardTestCase
+
+from picard.extension_points.script_functions import script_function
+from picard.plugin import ExtensionPoint
+from picard.script import (
+    ScriptFunctionDocError,
+    script_function_documentation,
+    script_function_documentation_all,
+)
+
+
+try:
+    from markdown import markdown  # type: ignore[unresolved-import]
+except ImportError:
+    markdown = None
+
+
+class ScriptFunctionDocumentationTest(PicardTestCase):
+    @staticmethod
+    def assertStartswith(text, expect):
+        if not text.startswith(expect):
+            raise AssertionError("do not start with %r but with %r" % (expect, text[: len(expect)]))
+
+    @staticmethod
+    def assertEndswith(text, expect):
+        if not text.endswith(expect):
+            raise AssertionError("do not end with %r but with %r" % (expect, text[-len(expect) :]))
+
+    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
+    def test_script_function_documentation_nodoc(self):
+        """test script_function_documentation() with a function without documentation"""
+
+        @script_function()
+        def func_nodocfunc(parser):
+            return ""
+
+        doc = script_function_documentation('nodocfunc', 'markdown')
+        self.assertEqual(doc, '`$nodocfunc()`')
+
+    @unittest.skipUnless(markdown, "markdown module missing")
+    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
+    def test_script_function_documentation_nodoc_html(self):
+        """test script_function_documentation() with a function without documentation"""
+
+        @script_function()
+        def func_nodocfunc(parser):
+            return ""
+
+        doc = script_function_documentation('nodocfunc', 'html')
+        self.assertEqual(doc, '<p><code>$nodocfunc()</code></p>')
+
+    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
+    def test_script_function_documentation(self):
+        """test script_function_documentation() with a function with documentation"""
+        # the documentation used to test includes backquotes
+        expected_doc = '`$somefunc()`\n\nsome doc'
+
+        @script_function(documentation='some doc')
+        def func_somefunc(parser):
+            return "x"
+
+        doc = script_function_documentation('somefunc', 'markdown')
+        self.assertEqual(doc, expected_doc)
+        areg = r"^no such documentation format: unknownformat"
+        with self.assertRaisesRegex(ScriptFunctionDocError, areg):
+            script_function_documentation('somefunc', 'unknownformat')
+
+    @unittest.skipUnless(markdown, "markdown module missing")
+    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
+    def test_script_function_documentation_html(self):
+        """test script_function_documentation() with a function with documentation"""
+        # get html code as generated by markdown
+        pre, post = markdown('`XXX`').split('XXX')
+
+        # the documentation used to test includes backquotes
+        expected_doc = '<p><code>$somefunc()</code></p>\n<p>some doc</p>'
+
+        @script_function(documentation='some doc')
+        def func_somefunc(parser):
+            return "x"
+
+        doc = script_function_documentation('somefunc', 'html')
+        self.assertEqual(doc, expected_doc)
+
+    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
+    def test_script_function_documentation_unknown_function(self):
+        """test script_function_documentation() with an unknown function"""
+        areg = r"^no such function: unknownfunc"
+        with self.assertRaisesRegex(ScriptFunctionDocError, areg):
+            script_function_documentation('unknownfunc', 'html')
+
+    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
+    def test_script_function_documentation_all(self):
+        """test script_function_documentation_all() with markdown format"""
+
+        @script_function(documentation='somedoc2')
+        def func_somefunc2(parser):
+            return "x"
+
+        @script_function(documentation='somedoc1')
+        def func_somefunc1(parser):
+            return "x"
+
+        docall = script_function_documentation_all()
+        self.assertEqual(docall, '`$somefunc1()`\n\nsomedoc1\n`$somefunc2()`\n\nsomedoc2')
+
+    @unittest.skipUnless(markdown, "markdown module missing")
+    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
+    def test_script_function_documentation_all_html(self):
+        """test script_function_documentation_all() with html format"""
+        # get html code as generated by markdown
+        pre, post = markdown('XXX').split('XXX')
+
+        @script_function(documentation='somedoc')
+        def func_somefunc(parser):
+            return "x"
+
+        def postprocessor(data, function):
+            return 'w' + data + function.name + 'y'
+
+        docall = script_function_documentation_all(
+            fmt='html',
+            pre='<div id="test">',
+            post="</div>\n",
+            postprocessor=postprocessor,
+        )
+
+        self.assertStartswith(docall, '<div id="test">w' + pre)
+        self.assertEndswith(docall, post + 'somefuncy</div>\n')
