+++ test/script/test_parser.py
@@ -31,7 +31,6 @@
 import datetime
 from inspect import getfullargspec
 import re
-import unittest
 from unittest import mock
 from unittest.mock import (
     MagicMock,
@@ -54,6 +53,7 @@ from picard.extension_points.script_functions import (
 from picard.metadata import (
     MULTI_VALUED_JOINER,
     Metadata,
+    MultiMetadataProxy,
 )
 from picard.plugin import ExtensionPoint
 from picard.script import (
@@ -62,14 +62,11 @@ from picard.script import (
     ScriptError,
     ScriptExpression,
     ScriptFunction,
-    ScriptFunctionDocError,
     ScriptParser,
     ScriptRuntimeError,
     ScriptSyntaxError,
     ScriptUnicodeError,
     ScriptUnknownFunction,
-    script_function_documentation,
-    script_function_documentation_all,
 )
 
 
@@ -275,118 +272,6 @@ class ScriptParserTest(PicardTestCase):
         self._assert_function_signature('$my_func(…)', 'my_func', lambda p, *args: None)
         self._assert_function_signature('$my_func(foo,…)', 'my_func', lambda p, foo, *args: None)
         self._assert_function_signature('$my_func(foo[,bar,…])', 'my_func', lambda p, foo, bar=None, *args: None)
-
-    @staticmethod
-    def assertStartswith(text, expect):
-        if not text.startswith(expect):
-            raise AssertionError("do not start with %r but with %r" % (expect, text[: len(expect)]))
-
-    @staticmethod
-    def assertEndswith(text, expect):
-        if not text.endswith(expect):
-            raise AssertionError("do not end with %r but with %r" % (expect, text[-len(expect) :]))
-
-    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
-    def test_script_function_documentation_nodoc(self):
-        """test script_function_documentation() with a function without documentation"""
-
-        @script_function()
-        def func_nodocfunc(parser):
-            return ""
-
-        doc = script_function_documentation('nodocfunc', 'markdown')
-        self.assertEqual(doc, '`$nodocfunc()`')
-
-    @unittest.skipUnless(markdown, "markdown module missing")
-    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
-    def test_script_function_documentation_nodoc_html(self):
-        """test script_function_documentation() with a function without documentation"""
-
-        @script_function()
-        def func_nodocfunc(parser):
-            return ""
-
-        doc = script_function_documentation('nodocfunc', 'html')
-        self.assertEqual(doc, '<p><code>$nodocfunc()</code></p>')
-
-    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
-    def test_script_function_documentation(self):
-        """test script_function_documentation() with a function with documentation"""
-        # the documentation used to test includes backquotes
-        expected_doc = '`$somefunc()`\n\nsome doc'
-
-        @script_function(documentation='some doc')
-        def func_somefunc(parser):
-            return "x"
-
-        doc = script_function_documentation('somefunc', 'markdown')
-        self.assertEqual(doc, expected_doc)
-        areg = r"^no such documentation format: unknownformat"
-        with self.assertRaisesRegex(ScriptFunctionDocError, areg):
-            script_function_documentation('somefunc', 'unknownformat')
-
-    @unittest.skipUnless(markdown, "markdown module missing")
-    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
-    def test_script_function_documentation_html(self):
-        """test script_function_documentation() with a function with documentation"""
-        # get html code as generated by markdown
-        pre, post = markdown('`XXX`').split('XXX')
-
-        # the documentation used to test includes backquotes
-        expected_doc = '<p><code>$somefunc()</code></p>\n<p>some doc</p>'
-
-        @script_function(documentation='some doc')
-        def func_somefunc(parser):
-            return "x"
-
-        doc = script_function_documentation('somefunc', 'html')
-        self.assertEqual(doc, expected_doc)
-
-    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
-    def test_script_function_documentation_unknown_function(self):
-        """test script_function_documentation() with an unknown function"""
-        areg = r"^no such function: unknownfunc"
-        with self.assertRaisesRegex(ScriptFunctionDocError, areg):
-            script_function_documentation('unknownfunc', 'html')
-
-    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
-    def test_script_function_documentation_all(self):
-        """test script_function_documentation_all() with markdown format"""
-
-        @script_function(documentation='somedoc2')
-        def func_somefunc2(parser):
-            return "x"
-
-        @script_function(documentation='somedoc1')
-        def func_somefunc1(parser):
-            return "x"
-
-        docall = script_function_documentation_all()
-        self.assertEqual(docall, '`$somefunc1()`\n\nsomedoc1\n`$somefunc2()`\n\nsomedoc2')
-
-    @unittest.skipUnless(markdown, "markdown module missing")
-    @patch('picard.extension_points.script_functions.ext_point_script_functions', ExtensionPoint(label='test_script'))
-    def test_script_function_documentation_all_html(self):
-        """test script_function_documentation_all() with html format"""
-        # get html code as generated by markdown
-        pre, post = markdown('XXX').split('XXX')
-
-        @script_function(documentation='somedoc')
-        def func_somefunc(parser):
-            return "x"
-
-        def postprocessor(data, function):
-            return 'w' + data + function.name + 'y'
-
-        docall = script_function_documentation_all(
-            fmt='html',
-            pre='<div id="test">',
-            post="</div>\n",
-            postprocessor=postprocessor,
-        )
-
-        self.assertStartswith(docall, '<div id="test">w' + pre)
-        self.assertEndswith(docall, post + 'somefuncy</div>\n')
 
     def test_unknown_function(self):
         areg = r"^\d+:\d+:\$unknownfunction: Unknown function '\$unknownfunction'"
@@ -2407,3 +2292,143 @@ class ScriptParserTest(PicardTestCase):
 
         # Tests with invalid number of arguments
         self._assert_arg_count_errors('max', 'at least 2', ("$max()", "$max(text)"))
+
+
+class TestGetNewFunction(PicardTestCase):
+    def setUp(self):
+        super().setUp()
+        self.set_config_values({'enabled_plugins': ''})
+        self.parser = ScriptParser()
+        ScriptParser._cache = {}
+
+    def test_get_new_with_proxy_tag_in_track_metadata(self):
+        """$get_new returns value when tag exists in track/MB metadata."""
+        track_metadata = Metadata()
+        track_metadata['composer'] = 'MB Composer'
+
+        file_metadata = Metadata()
+
+        proxy = MultiMetadataProxy(Metadata(track_metadata), file_metadata)
+        result = self.parser.eval("$get_new(composer)", proxy)
+        self.assertEqual(result, "MB Composer")
+
+    def test_get_new_with_proxy_tag_only_in_file(self):
+        """$get_new returns empty when tag only exists in file metadata."""
+        track_metadata = Metadata()
+
+        file_metadata = Metadata()
+        file_metadata['composer'] = 'Old File Composer'
+
+        proxy = MultiMetadataProxy(Metadata(track_metadata), file_metadata)
+        result = self.parser.eval("$get_new(composer)", proxy)
+        self.assertEqual(result, "")
+
+    def test_get_new_with_proxy_tag_in_both(self):
+        """$get_new returns track value when tag exists in both."""
+        track_metadata = Metadata()
+        track_metadata['composer'] = 'MB Composer'
+
+        file_metadata = Metadata()
+        file_metadata['composer'] = 'Old File Composer'
+
+        proxy = MultiMetadataProxy(Metadata(track_metadata), file_metadata)
+        result = self.parser.eval("$get_new(composer)", proxy)
+        self.assertEqual(result, "MB Composer")
+
+    def test_get_new_with_proxy_tag_nowhere(self):
+        """$get_new returns empty when tag doesn't exist anywhere."""
+        track_metadata = Metadata()
+        file_metadata = Metadata()
+
+        proxy = MultiMetadataProxy(Metadata(track_metadata), file_metadata)
+        result = self.parser.eval("$get_new(composer)", proxy)
+        self.assertEqual(result, "")
+
+    def test_get_new_with_plain_metadata(self):
+        """$get_new on plain Metadata returns value (same as %name%)."""
+        metadata = Metadata()
+        metadata['composer'] = 'Some Composer'
+
+        result = self.parser.eval("$get_new(composer)", metadata)
+        self.assertEqual(result, "Some Composer")
+
+    def test_get_new_used_in_if_condition(self):
+        """$get_new works in $if for the user's use case."""
+        track_metadata = Metadata()
+        track_metadata['lyricist'] = 'Some Lyricist'
+
+        file_metadata = Metadata()
+        file_metadata['composer'] = 'Old File Composer'
+
+        metadata = Metadata(track_metadata)
+        proxy = MultiMetadataProxy(metadata, file_metadata)
+
+        script = '$if($get_new(composer),,$set(composer,$if2(%lyricist%,%writer%,%arranger%)))'
+        self.parser.eval(script, proxy)
+
+        # Script fires because $get_new(composer) is empty
+        self.assertEqual(metadata['composer'], 'Some Lyricist')
+
+
+class TestGetOriginalFunction(PicardTestCase):
+    def setUp(self):
+        super().setUp()
+        self.set_config_values({'enabled_plugins': ''})
+        self.parser = ScriptParser()
+        ScriptParser._cache = {}
+
+    def test_get_original_with_proxy_tag_in_file(self):
+        """$get_original returns file's value."""
+        track_metadata = Metadata()
+        track_metadata['composer'] = 'MB Composer'
+
+        file_metadata = Metadata()
+        file_metadata['composer'] = 'Old File Composer'
+
+        proxy = MultiMetadataProxy(Metadata(track_metadata), file_metadata)
+        result = self.parser.eval("$get_original(composer)", proxy)
+        self.assertEqual(result, "Old File Composer")
+
+    def test_get_original_with_proxy_tag_only_in_track(self):
+        """$get_original returns empty when tag only exists in track metadata."""
+        track_metadata = Metadata()
+        track_metadata['composer'] = 'MB Composer'
+
+        file_metadata = Metadata()
+
+        proxy = MultiMetadataProxy(Metadata(track_metadata), file_metadata)
+        result = self.parser.eval("$get_original(composer)", proxy)
+        self.assertEqual(result, "")
+
+    def test_get_original_with_proxy_tag_nowhere(self):
+        """$get_original returns empty when tag doesn't exist anywhere."""
+        track_metadata = Metadata()
+        file_metadata = Metadata()
+
+        proxy = MultiMetadataProxy(Metadata(track_metadata), file_metadata)
+        result = self.parser.eval("$get_original(composer)", proxy)
+        self.assertEqual(result, "")
+
+    def test_get_original_with_plain_metadata(self):
+        """$get_original on plain Metadata returns value (same as %name%)."""
+        metadata = Metadata()
+        metadata['composer'] = 'Some Composer'
+
+        result = self.parser.eval("$get_original(composer)", metadata)
+        self.assertEqual(result, "Some Composer")
+
+    def test_get_original_use_in_script(self):
+        """$get_original can be used to reference file's original value."""
+        track_metadata = Metadata()
+        track_metadata['artist'] = 'MB Artist'
+
+        file_metadata = Metadata()
+        file_metadata['artist'] = 'File Artist'
+
+        metadata = Metadata(track_metadata)
+        proxy = MultiMetadataProxy(metadata, file_metadata)
+
+        script = '$set(comment,Original: $get_original(artist))'
+        self.parser.eval(script, proxy)
+
+        self.assertEqual(metadata['comment'], 'Original: File Artist')
